Fix settings rewriting existing settings
This only occurs if the ID could not be found & the assets path is the same as the default path. It will still override the asset if an asset of another type exists in it's place.

This also simultaneously fixes A bug where the default canvas is set to null upon the package first being installed.